soc-change

+++ include/oledDisplay.hpp
@@ -26,7 +26,7 @@ extern Adafruit_SSD1306 oled;
 
 /* ----------- Test data ----------- */
 static int speed[] = {10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20};
-static int state_of_charge[] = {100, 99, 98, 97, 96, 95, 94, 93, 92, 91, 90};
+static float state_of_charge[] = {100, 99, 98, 97, 96, 95, 94, 93, 92, 91, 90};
 
 /* ----------- Function declarations ----------- */
 void init_oled_display();
